worker/requirements.txt
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
feedparser==6.0.10
requests==2.31.0
beautifulsoup4==4.12.2
arxiv==1.4.8
python-dateutil==2.8.2
nltk==3.8.1
scikit-learn==1.3.2
numpy==1.26.2
